+++ src/array/codec/array_to_array/transpose/transpose_codec.rs
@@ -139,21 +139,16 @@ impl ArrayCodecTraits for TransposeCodec {
                 decoded_representation.size(),
             ));
         }
-
+        let len = decoded_value.len();
         let order_encode =
             calculate_order_encode(&self.order, decoded_representation.shape().len());
         transpose_array(
             &order_encode,
             &decoded_representation.shape_u64(),
             decoded_representation.element_size(),
-            &decoded_value,
+            decoded_value,
         )
-        .map_err(|_| {
-            CodecError::UnexpectedChunkDecodedSize(
-                decoded_value.len(),
-                decoded_representation.size(),
-            )
-        })
+        .map_err(|_| CodecError::UnexpectedChunkDecodedSize(len, decoded_representation.size()))
     }
 
     fn decode(
@@ -165,17 +160,13 @@ impl ArrayCodecTraits for TransposeCodec {
         let order_decode =
             calculate_order_decode(&self.order, decoded_representation.shape().len());
         let transposed_shape = permute(&decoded_representation.shape_u64(), &self.order);
+        let len = encoded_value.len();
         transpose_array(
             &order_decode,
             &transposed_shape,
             decoded_representation.element_size(),
-            &encoded_value,
+            encoded_value,
         )
-        .map_err(|_| {
-            CodecError::UnexpectedChunkDecodedSize(
-                encoded_value.len(),
-                decoded_representation.size(),
-            )
-        })
+        .map_err(|_| CodecError::UnexpectedChunkDecodedSize(len, decoded_representation.size()))
     }
 }